# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/degrees": {
      "post": {
        "description": "Creates a new degree record.",
        "operationId": "DegreesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDegreeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created degree.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Degree"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new degree",
        "tags": [
          "degrees"
        ]
      },
      "get": {
        "description": "Returns a list of all degrees.",
        "operationId": "DegreesController_findAll",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facultyId",
            "required": false,
            "in": "query",
            "description": "Filter by faculty id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "description": "Filter by university id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of degrees.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Degree"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve all degrees",
        "tags": [
          "degrees"
        ]
      }
    },
    "/degrees/bulk": {
      "post": {
        "description": "Creates multiple degree records in a single operation.",
        "operationId": "DegreesController_createMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManyDegreesInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The number of degrees created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "description": "Number of degrees created",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create multiple degrees",
        "tags": [
          "degrees"
        ]
      }
    },
    "/degrees/{id}": {
      "get": {
        "description": "Returns a single degree by its ID.",
        "operationId": "DegreesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The degree with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Degree"
                }
              }
            }
          },
          "404": {
            "description": "Degree not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve a degree by ID",
        "tags": [
          "degrees"
        ]
      },
      "put": {
        "description": "Updates an existing degree record.",
        "operationId": "DegreesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated degree.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Degree"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Degree not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a degree by ID",
        "tags": [
          "degrees"
        ]
      },
      "delete": {
        "description": "Deletes a degree record by its ID.",
        "operationId": "DegreesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Degree successfully deleted."
          },
          "404": {
            "description": "Degree not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a degree by ID",
        "tags": [
          "degrees"
        ]
      }
    },
    "/degrees/{id}/faculty": {
      "post": {
        "description": "Sets or updates the faculty associated with a degree.",
        "operationId": "DegreesController_setFaculty",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The degree with the updated faculty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Degree"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Degree or Faculty not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Set faculty for a degree",
        "tags": [
          "degrees"
        ]
      }
    },
    "/degrees/{id}/courses": {
      "post": {
        "description": "Adds a course to the specified degree.",
        "operationId": "DegreesController_addCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The degree with the added course.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Degree"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Degree or Course not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Add a course to a degree",
        "tags": [
          "degrees"
        ]
      }
    },
    "/degrees/{id}/summary": {
      "get": {
        "description": "Returns a human-readable plain text summary for the specified degree.",
        "operationId": "DegreesController_getSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plain text summary of the degree",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Degree not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get degree summary",
        "tags": [
          "degrees"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumina API",
    "description": "API documentation for Lumina server",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Translation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "en_text": {
            "type": "string"
          },
          "he_text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "en_text",
          "he_text"
        ]
      },
      "Faculty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "description": {
            "$ref": "#/components/schemas/Translation"
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "degrees"
        ]
      },
      "Degree": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "facultyId": {
            "type": "string",
            "nullable": true
          },
          "faculty": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Faculty"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "courses",
          "facultyId",
          "faculty"
        ]
      },
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "courses",
          "degrees"
        ]
      },
      "RelationshipMetadataKey": {
        "type": "string",
        "enum": [
          "REASON",
          "TYPE",
          "DESCRIPTION"
        ]
      },
      "RelationshipMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RelationshipMetadataKey"
              }
            ]
          },
          "value": {
            "type": "string"
          },
          "blockRelationshipId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "value",
          "blockRelationshipId"
        ]
      },
      "BlockRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteId": {
            "type": "string"
          },
          "postrequisiteId": {
            "type": "string"
          },
          "prerequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "postrequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "metadata": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationshipMetadata"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteId",
          "postrequisiteId",
          "prerequisite",
          "postrequisite",
          "metadata"
        ]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteFor": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "postrequisiteOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteFor",
          "postrequisiteOf",
          "modules"
        ]
      },
      "QuestionValidationStatus": {
        "type": "string",
        "enum": [
          "ai_generated",
          "in_manual_review",
          "approved",
          "rejected"
        ]
      },
      "SelectAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "translationId",
          "text",
          "answerId"
        ]
      },
      "UnitAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "unit",
          "answerId"
        ]
      },
      "NumberAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "answerId"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "SelectAnswer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectAnswer"
            }
          },
          "UnitAnswer": {
            "$ref": "#/components/schemas/UnitAnswer"
          },
          "NumberAnswer": {
            "$ref": "#/components/schemas/NumberAnswer"
          }
        },
        "required": [
          "id",
          "questionId"
        ]
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "selection",
          "value",
          "void"
        ]
      },
      "QuestionPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "partQuestionId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "questionId",
          "partQuestionId",
          "order"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "validationStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionValidationStatus"
              }
            ]
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "Answer": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              }
            ]
          },
          "Modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Parts": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          },
          "PartOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          }
        },
        "required": [
          "id",
          "validationStatus",
          "translationId",
          "text",
          "Answer",
          "type",
          "Modules",
          "Parts",
          "PartOf"
        ]
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "translationId": {
            "type": "string"
          },
          "subModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "parentModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "questions": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        },
        "required": [
          "id",
          "name",
          "translationId",
          "subModules",
          "parentModules",
          "Block",
          "questions"
        ]
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "modules",
          "publishedAt",
          "Block"
        ]
      },
...
      "CreateDegreeInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Degree name"
          },
          "universityId": {
            "type": "string",
            "description": "Owning university id",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "universityId"
        ]
      },
      "CreateManyDegreesInput": {
        "type": "object",
        "properties": {
          "degrees": {
            "description": "Array of degree data to create",
            "example": [
              {
                "name": "Bachelor of Science",
                "universityId": "uuid1"
              },
              {
                "name": "Master of Arts",
                "universityId": "uuid2"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDegreeInput"
            }
          }
        },
        "required": [
          "degrees"
        ]
      },
...
    }
  }
}